backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# HTTP Clients
httpx==0.25.1
requests==2.31.0

# LLM SDKs (install the one you need)
groq==0.11.0
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0

# CORS and Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Async Support
aiofiles==23.2.1
asyncio==3.4.3

# Development Tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and Logging (optional)
loguru==0.7.2
prometheus-client==0.19.0
